+++ utils.go
@@ -16,10 +16,25 @@ package centraldogma
 
 import (
 	"fmt"
+	"math"
+	"math/rand"
 	"net/url"
 	"strconv"
 	"strings"
+	"time"
 )
+
+const (
+	delayOnSuccess = 1 * time.Second
+	minInterval    = delayOnSuccess * 2
+	maxInterval    = 1 * time.Minute
+	jitterRate     = 0.2
+	maxInt63       = int64(^uint64(0) >> 1)
+)
+
+func init() {
+	rand.Seed(time.Now().UTC().UnixNano())
+}
 
 func setURLValue(v *url.Values, key, value string) {
 	if len(value) > 0 {
@@ -70,4 +85,60 @@ func setJSONPaths(v *url.Values, query *Query) (err error) {
 		}
 	}
 	return
+}
+
+func nextDelay(numAttemptsSoFar int) time.Duration {
+	var nextDelay time.Duration
+	if numAttemptsSoFar == 1 {
+		nextDelay = minInterval
+	} else {
+		calculatedDelay := saturatedMultiply(minInterval, math.Pow(2.0, float64(numAttemptsSoFar-1)))
+		if calculatedDelay > maxInterval {
+			nextDelay = maxInterval
+		} else {
+			nextDelay = calculatedDelay
+		}
+	}
+	minJitter := int64(float64(nextDelay) * (1 - jitterRate))
+	maxJitter := int64(float64(nextDelay) * (1 + jitterRate))
+	bound := maxJitter - minJitter + 1
+	random := random(bound)
+	result := saturatedAdd(minJitter, random)
+	if result < 0 {
+		result = 0
+	}
+	return time.Duration(result)
+}
+
+func saturatedMultiply(left time.Duration, right float64) time.Duration {
+	result := float64(left) * right
+	if result > float64(maxInt63) {
+		return time.Duration(maxInt63)
+	}
+	return time.Duration(result)
+}
+
+func random(bound int64) int64 {
+	mask := bound - 1
+	result := rand.Int63()
+
+	if bound&mask == 0 {
+		// power of two
+		result &= mask
+	} else { // reject over-represented candidates
+		for u := result >> 1; u+mask-result < 0; u = rand.Int63() >> 1 {
+			result = u % bound
+		}
+	}
+	return result
+}
+
+// This code is from Guava library.
+func saturatedAdd(a, b int64) int64 {
+	naiveSum := a + b
+	if a^b < 0 || a^naiveSum >= 0 {
+		// If a and b have different signs or a has the same sign as the result then there was no overflow.
+		return naiveSum
+	}
+	return maxInt63 + ((naiveSum >> (64 - 1)) ^ 1)
 }
